role based product

+++ src/controllers/product.controller.js
@@ -6,7 +6,8 @@ const createProduct = async (req, res, next) => {
       ...req.body,
       createdBy: req.user.id,
     });
-    res.status(201).json(product);
+
+    return res.status(201).json(product);
   } catch (error) {
     next(error);
   }
@@ -14,11 +15,11 @@ const createProduct = async (req, res, next) => {
 
 const getProducts = async (req, res, next) => {
   try {
-    const products = await Product.find().populate(
-      "createdBy",
-      "name email"
-    );
-    res.json(products);
+    const products = await Product.find({
+      createdBy: req.user.id,
+    }).populate("createdBy", "name email");
+    
+    return res.status(200).json(products);
   } catch (error) {
     next(error);
   }
@@ -29,38 +30,55 @@ const updateProduct = async (req, res, next) => {
     const { id } = req.params;
 
     const product = await Product.findById(id);
-    console.log("FOUND PRODUCT:", product);
-
     if (!product) {
       return res.status(404).json({ message: "Product not found" });
     }
 
-    const updatedProduct = await Product.findByIdAndUpdate(
-      id,
-      req.body,
-      { new: true, runValidators: true, }
-    );
+    if (product.createdBy.toString() !== req.user.id) {
+      return res.status(403).json({
+        message: "You are not allowed to update this product",
+      });
+    }
 
-    res.json(updatedProduct);
+    const updatedProduct = await Product.findByIdAndUpdate(id, req.body, {
+      new: true,
+      runValidators: true,
+    });
+
+    return res.status(200).json(updatedProduct);
   } catch (error) {
     next(error);
   }
 };
-
 
 const deleteProduct = async (req, res, next) => {
   try {
     const { id } = req.params;
 
     const product = await Product.findById(id);
-    if (!product){
+    if (!product) {
       return res.status(404).json({ message: "Product not found" });
     }
+
+    if (product.createdBy.toString() !== req.user.id) {
+      return res.status(403).json({
+        message: "You are not allowed to delete this product",
+      });
+    }
+
     await product.deleteOne();
-    res.json({ message: "Product deleted successfully" });
+
+    return res.status(200).json({
+      message: "Product deleted successfully",
+    });
   } catch (error) {
     next(error);
   }
 };
 
-module.exports = { createProduct, getProducts, updateProduct, deleteProduct};
+module.exports = {
+  createProduct,
+  getProducts,
+  updateProduct,
+  deleteProduct,
+};
